requirements.txt
transformers
ipykernel
jupyter
tensorflow
torch
mysql-connector-python 
pandas
nltk
matplotlib
wordcloud
praw
psaw
scipy
scikit-learn